add: outsourced drawing into its own function
